+++ newrelic/common/object_wrapper.py
@@ -7,6 +7,7 @@ make use of when doing monkey patching.
 
 import sys
 import inspect
+import warnings
 
 from newrelic.packages import six
 
@@ -151,10 +152,12 @@ class ObjectWrapper(_ObjectWrapperBase, _FunctionWrapperBase):
         super(ObjectWrapper, self).__init__(wrapped, instance, wrapper,
                 binding=binding)
 
-# The wrap_callable() alias needs to be deprecated and usage of it removed.
 
-
-wrap_callable = FunctionWrapper
+def wrap_callable(*args, **kwargs):
+    warnings.warn('Internal API change. Use FunctionWrapper '
+            'instead of wrap_callable.', DeprecationWarning,
+            stacklevel=2)
+    return FunctionWrapper(*args, **kwargs)
 
 # Helper functions for performing monkey patching.